Move Root.jsx up level up
