+++ app/Root.jsx
@@ -1,0 +1,18 @@
+// @flow
+import React, { Component } from 'react'
+import { Provider } from 'react-redux'
+import Routes from './Routes'
+import { BrowserRouter } from 'react-router-dom'
+
+type Props = {
+  store: Object,
+}
+
+const Root = ({ store }: Props) =>
+  <Provider store={store}>
+    <BrowserRouter>
+      <Routes />
+    </BrowserRouter>
+  </Provider>
+
+export default Root
